CityData: Corrected building maintenance deficit spending handling. BldQue: Improved handling of multiple build queue modifications within a turn.
casts, type corrections, scope corrections.

git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@504 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/gs/database/StrDB.cpp
@@ -3,7 +3,7 @@
 // Project      : Call To Power 2
 // File type    : C++ source
 // Description  : String database
-// Id           : $Id:$
+// Id           : $Id$
 //
 //----------------------------------------------------------------------------
 //
@@ -275,7 +275,7 @@ bool StringDB::GetText(MBCHAR const * get_id, MBCHAR ** new_text) const
 MBCHAR * StringDB::GetIdStr(StringId const & index) const
 { 
 	Assert(0 <= index);
-	Assert(index < m_all.size());
+	Assert(static_cast<size_t>(index) < m_all.size());
 	return (index < 0) || (index >= static_cast<sint32>(m_all.size())) ? NULL : m_all[index]->m_id;
 }
